+++ baybe/recommenders/pure/bayesian/base.py
@@ -19,38 +19,17 @@ from baybe.exceptions import (
     IncompatibleAcquisitionFunctionError,
 )
 from baybe.objectives.base import Objective
-from baybe.recommenders.pure.base import PureRecommender
+from baybe.recommenders.pure.surrogate import SurrogateRecommender
 from baybe.searchspace import SearchSpace
-from baybe.settings import Settings
-from baybe.surrogates import GaussianProcessSurrogate
-from baybe.surrogates.base import (
-    Surrogate,
-    SurrogateProtocol,
-)
 from baybe.symmetries.base import Symmetry
-from baybe.utils.validation import preprocess_dataframe, validate_object_names
 
 if TYPE_CHECKING:
     from botorch.acquisition import AcquisitionFunction as BoAcquisitionFunction
 
 
-def _autoreplicate(surrogate: SurrogateProtocol, /) -> SurrogateProtocol:
-    """Replicates single-output surrogate models and passes through everything else."""
-    if isinstance(surrogate, Surrogate) and not surrogate.supports_multi_output:
-        return surrogate.replicate()
-    return surrogate
-
-
 @define
-class BayesianRecommender(PureRecommender, ABC):
+class BayesianRecommender(SurrogateRecommender, ABC):
     """An abstract class for Bayesian Recommenders."""
-
-    _surrogate_model: SurrogateProtocol = field(
-        alias="surrogate_model",
-        factory=GaussianProcessSurrogate,
-        converter=_autoreplicate,
-    )
-    """The surrogate model."""
 
     acquisition_function: AcquisitionFunction | None = field(
         default=None, converter=optional(convert_acqf)
@@ -83,16 +62,20 @@ class BayesianRecommender(PureRecommender, ABC):
             return qLogNEHVI() if objective.is_multi_output else qLogEI()
         return self.acquisition_function
 
-    def get_surrogate(
+    @override
+    def _prepare_recommendation(
         self,
         searchspace: SearchSpace,
         objective: Objective,
         measurements: pd.DataFrame,
-    ) -> SurrogateProtocol:
-        """Get the trained surrogate model."""
-        # This fit applies internal caching and does not necessarily involve computation
-        self._surrogate_model.fit(searchspace, objective, measurements)
-        return self._surrogate_model
+        pending_experiments: pd.DataFrame | None,
+    ) -> None:
+        self._setup_botorch_acqf(
+            searchspace=searchspace,
+            objective=objective,
+            measurements=measurements,
+            pending_experiments=pending_experiments,
+        )
 
     def _setup_botorch_acqf(
         self,
@@ -149,48 +132,14 @@ class BayesianRecommender(PureRecommender, ABC):
         measurements: pd.DataFrame | None = None,
         pending_experiments: pd.DataFrame | None = None,
     ) -> pd.DataFrame:
-        if objective is None:
-            raise NotImplementedError(
-                f"Recommenders of type '{BayesianRecommender.__name__}' require "
-                f"that an objective is specified."
-            )
-
-        validate_object_names(searchspace.parameters + objective.targets)
-
-        # Experimental input validation
-        if (measurements is None) or measurements.empty:
-            raise NotImplementedError(
-                f"Recommenders of type '{BayesianRecommender.__name__}' do not support "
-                f"empty training data."
-            )
-
-        measurements = preprocess_dataframe(
-            measurements,
-            searchspace,
-            objective,
-            numerical_measurements_must_be_within_tolerance=False,
-        )
-
-        if pending_experiments is not None:
-            pending_experiments = preprocess_dataframe(
-                pending_experiments,
-                searchspace,
-                numerical_measurements_must_be_within_tolerance=False,
-            )
-
-        self._setup_botorch_acqf(
-            searchspace, objective, measurements, pending_experiments
-        )
-
         try:
-            with Settings(preprocess_dataframes=False):
-                return super().recommend(
-                    batch_size=batch_size,
-                    searchspace=searchspace,
-                    objective=objective,
-                    measurements=measurements,
-                    pending_experiments=pending_experiments,
-                )
+            return super().recommend(
+                batch_size=batch_size,
+                searchspace=searchspace,
+                objective=objective,
+                measurements=measurements,
+                pending_experiments=pending_experiments,
+            )
         except RuntimeError as ex:
             # Search spaces with continuous components are incompatible with surrogates
             # that do not support gradient computation